Separating web_crawler component
