+++ .gitignore
@@ -2,7 +2,6 @@ run.cmd
 #Decision Files
 Decision Files/*
 
-geckodriver.exe
 # Byte-compiled / optimized / DLL files
 __pycache__/
 *.py[cod]
@@ -15,7 +14,6 @@ __pycache__/
 .Python
 build/
 develop-eggs/
-dist/
 downloads/
 eggs/
 .eggs/
